+++ Program/income.h
@@ -8,7 +8,7 @@
 #include <dos.h>
 #include <limits.h>
 // Go to Menu1
-int Menu1();
+float Menu1();
 //Go to Menu2
 //void Menu2();
 //Go to Menu3
@@ -20,6 +20,11 @@ bool FindElement(int,int*,int);
 bool Valid_daymonth(int,int);
 
 struct list{
+	//inex: income or expense.
+	//type: type of item.
+	//amount: amount of money.
+	//name: name of item.
+	//detail: detail of item.
 	int type;
 	float amount;
 	char name[50], detail[100], date[15];
